refactor (#4) : 공통 응답 class global package로 분리
