+++ src/main/java/com/limito/limitedproduct/global/dto/response/ProductItemResponseInfo.java
@@ -1,0 +1,17 @@
+package com.limito.limitedproduct.global.dto.response;
+
+import java.util.UUID;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class ProductItemResponseInfo {
+	private UUID limitedProductItemId;
+	private String size;
+	private int price;
+	private int purchaseAmountLimit;
+	private int stock;
+	private boolean soldOut;
+}
