+++ src/main/java/com/limito/limitedproduct/global/dto/response/ProductResponseInfo.java
@@ -1,0 +1,16 @@
+package com.limito.limitedproduct.global.dto.response;
+
+import java.util.UUID;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class ProductResponseInfo {
+	private UUID limitedProductId;
+	private UUID categoryId;
+	private String name;
+	private Long sellerId;
+	private String brandName;
+}
